Move library test into tests folder
